+++ test/e2e/test_fast_xmlupload.py
@@ -39,32 +39,35 @@ class TestFastXmlUpload(unittest.TestCase):
             verbose=False,
             dump=False,
         )
-        shutil.copytree(cls.input_dir, cls.input_dir / "nested")
-        shutil.copytree(cls.input_dir / "nested", cls.input_dir / "nested/subfolder")
+        Path(cls.input_dir / "nested").mkdir()
+        Path(cls.input_dir / "nested/subfolder").mkdir()
+        shutil.copy(cls.input_dir / "test.jpg", cls.input_dir / "nested/test.jpg")
+        shutil.copy(cls.input_dir / "test.jpg", cls.input_dir / "nested/subfolder/test.jpg")
 
     @classmethod
     def tearDownClass(cls) -> None:
         """
-        Is executed after the methods of this class have all run through
+        Remove the testdata/bitstreams/nested folder and the testdata/preprocessed_files folder.
+        Is executed after the methods of this class have all run through.
         """
         shutil.rmtree(cls.input_dir / "nested")
         shutil.rmtree(cls.output_dir)
 
     def tearDown(self) -> None:
         """
-        Is executed after each test method
+        Delete all pickle files, all id2iri files and all "(un)processed_files.txt" files.
+        Is executed after each test method.
         """
         for pickle_file in list(Path().glob("*.pkl")):
             pickle_file.unlink()
 
-        id2iri_search_results = list(Path().glob("*id2iri_mapping.json"))
-        if len(id2iri_search_results) == 1:
-            id2iri_search_results[0].unlink()
+        for id2iri_file in list(Path().glob("*id2iri_mapping.json")):
+            id2iri_file.unlink()
 
         for txt_file in self.txt_files:
             Path(txt_file).unlink(missing_ok=True)
 
-    def test_fast_xmlupload(self) -> None:
+    def test_fast_xmlupload_without_batching(self) -> None:
         """
         Test if it is possible to call the 3 steps of the fast XML upload, one after the other.
         No detailed tests are done here, it is only tested if the 3 steps can be called.
@@ -79,11 +82,11 @@ class TestFastXmlUpload(unittest.TestCase):
         )
         self.assertTrue(success_process)
 
-        pickle_file = str(list(Path().glob("*.pkl"))[0])
+        pickle_files = list(Path().glob("*.pkl"))
+        self.assertEqual(len(pickle_files), 1)
 
-        print(f"test_fast_xmlupload: call upload_files() with pickle file {pickle_file}")
+        print("test_fast_xmlupload: call upload_files()")
         success_upload = upload_files(
-            pkl_file=pickle_file,
             dir_with_processed_files=self.output_dir,
             nthreads=4,
             user=self.user,
@@ -96,7 +99,6 @@ class TestFastXmlUpload(unittest.TestCase):
         print("test_fast_xmlupload: call fast_xmlupload()")
         success_fast_xmlupload = fast_xmlupload(
             xml_file=self.xml_file,
-            pkl_file=pickle_file,
             user=self.user,
             password=self.password,
             dsp_url=self.dsp_url,
@@ -104,21 +106,24 @@ class TestFastXmlUpload(unittest.TestCase):
         )
         self.assertTrue(success_fast_xmlupload)
 
-    def test_batch_size_of_process_files(self) -> None:
+    def test_fast_xmlupload_batching(self) -> None:
         """
         Test if the "batch_size" parameter of process_files() function works.
-        The test file contains 92 bitstreams, so a batch size of 40 should result in 3 batches.
+        The test file contains 34 bitstreams, so a batch size of 15 should result in 3 batches.
         The first 2 batches should exit with exit code 2 and success=True,
         the 3rd batch should exit with exit code 0 and success=True.
+
+        Afterwards, the uploading step and the fast xmlupload step should work with the 3 pickle files.
         """
 
         def action() -> bool:
+            print("test_fast_xmlupload_batching: call process_files() with batch size 15")
             return process_files(
                 input_dir=str(self.input_dir),
                 output_dir=self.output_dir,
                 xml_file=self.xml_file,
                 nthreads=None,
-                batch_size=40,
+                batch_size=15,
             )
 
         for i in range(2):
@@ -129,6 +134,30 @@ class TestFastXmlUpload(unittest.TestCase):
 
         success = action()
         self.assertTrue(success)
+
+        pickle_files = list(Path().glob("*.pkl"))
+        self.assertEqual(len(pickle_files), 3)
+
+        print("test_fast_xmlupload_batching: call upload_files()")
+        success_upload = upload_files(
+            dir_with_processed_files=self.output_dir,
+            nthreads=4,
+            user=self.user,
+            password=self.password,
+            dsp_url=self.dsp_url,
+            sipi_url=self.sipi_url,
+        )
+        self.assertTrue(success_upload)
+
+        print("test_fast_xmlupload_batching: call fast_xmlupload()")
+        success_fast_xmlupload = fast_xmlupload(
+            xml_file=self.xml_file,
+            user=self.user,
+            password=self.password,
+            dsp_url=self.dsp_url,
+            sipi_url=self.sipi_url,
+        )
+        self.assertTrue(success_fast_xmlupload)
 
 
 if __name__ == "__main__":